+++ Application/RA.Utilities.Feature/Abstractions/IMediator.cs
@@ -21,7 +21,7 @@ public interface IMediator
     /// <summary>
     /// Sends a request with a response.
     /// </summary>
-    /// <typeparam name="TRequest">The type of the reques.</typeparam>
+    /// <typeparam name="TRequest">The type of the request.</typeparam>
     /// <typeparam name="TResponse">The type of the response.</typeparam>
     /// <param name="request">The request to send.</param>
     /// <param name="cancellationToken">The cancellation token.</param>